Rework celery request
Request are now fully parallel (with the exception of paginablejson with more
than one page). Big architecture issue but hopefully do not appear to often

On Genymotion, with 80 repos and 23 open PR
 Before: 19s
 After: 9s

Also add a --beat option to celeryd that is supposed to purge the sqlite
file regulary.

It is possible than we can improve the perf using something else than sqlite.
It seems that all the jobs cannot really be launched simustaneously.